+++ packaging/nix/default.nix
@@ -8,7 +8,7 @@
 
 stdenv.mkDerivation rec {
   pname = "retrace";
-  version = "2.3.0";
+  version = "2.3.1";
 
   src = ./.;